refactor(workbench): resolve circular import cycles by moving model classes to separate files
