+++ projects/scion/workbench/src/lib/spec/view-part.model.spec.ts
@@ -9,9 +9,10 @@
  */
 
 import { Observable, Subject } from 'rxjs';
-import { WbBeforeDestroy, WorkbenchView } from '../workbench.model';
+import { WbBeforeDestroy } from '../workbench.model';
 import { Component, OnDestroy } from '@angular/core';
 import { takeUntil } from 'rxjs/operators';
+import { WorkbenchView } from '../view/workbench-view.model';
 
 @Component({})
 export abstract class AbstractSpecViewComponent implements OnDestroy, WbBeforeDestroy {
