+++ projects/scion/workbench/src/lib/view-part/view-list/view-list.component.ts
@@ -14,7 +14,7 @@ import { animate, AnimationBuilder, style } from '@angular/animations';
 import { fromEvent, Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { SciDimension } from '@scion/toolkit/dimension';
-import { InternalWorkbenchViewPart } from '../../workbench.model';
+import { ɵWorkbenchViewPart } from '../ɵworkbench-view-part.model';
 
 @Component({
   selector: 'wb-view-list',
@@ -33,7 +33,7 @@ export class ViewListComponent implements OnInit, OnDestroy {
   public hiddenViewTabs: string[] = [];
 
   constructor(private _host: ElementRef<HTMLElement>,
-              private _viewPart: InternalWorkbenchViewPart,
+              private _viewPart: ɵWorkbenchViewPart,
               private _overlayRef: OverlayRef,
               private _animationBuilder: AnimationBuilder) {
     this.installHiddenViewTabsListener();
